removing s3 file before push
