+++ .github/workflows/deploy-stage.yml
@@ -1,5 +1,4 @@
 name: BetterFinco stage Frontend deploy
-1
 on:
     push:
         branches: [staging-betterfinco]
@@ -20,6 +19,13 @@ jobs:
               run: npm install
             - name: Build Application
               run: npm run build:dev
+            - name: Clean S3 Bucket
+              run: |
+                  aws s3 rm s3://${{ vars.AWS_BUCKET }} --recursive
+              env:
+                  AWS_ACCESS_KEY_ID: ${{ secrets.AWS_ACCESS_KEY_ID }}
+                  AWS_SECRET_ACCESS_KEY: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+                  AWS_REGION: ${{ vars.AWS_REGION }}
             - uses: jakejarvis/s3-sync-action@master
               name: Upload App to S3 Bucket
               with:
