+++ src/app/account-transfers/make-account-transfers/make-account-transfers.component.ts
@@ -18,7 +18,7 @@ import { Dates } from 'app/core/utils/dates';
 import { MatInput } from '@angular/material/input';
 
 /** Environment Configuration */
-import { environment } from 'environments/environment';
+import { environment } from 'environments/environment.dev';
 
 /**
  * Create account transfers
