+++ src/app/clients/clients.component.ts
@@ -6,7 +6,7 @@ import { MatSort, Sort } from '@angular/material/sort';
 import { MatTableDataSource } from '@angular/material/table';
 
 /** Custom Services */
-import { environment } from 'environments/environment';
+import { environment } from 'environments/environment.dev';
 import { ClientsService } from './clients.service';
 
 @Component({
