+++ src/app/core/authentication/authentication.interceptor.ts
@@ -6,7 +6,7 @@ import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest } from '@angular/c
 import { Observable } from 'rxjs';
 
 /** Custom Imports */
-import { environment } from '../../../environments/environment';
+import { environment } from '../../../environments/environment.dev';
 import { SettingsService } from 'app/settings/settings.service';
 
 /** Http request (default) options headers. */
