+++ src/app/core/utils/passwords-utility.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { AbstractControl, ValidationErrors, ValidatorFn, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
 import { passwordValidator } from './password.validator';
-import { environment } from 'environments/environment';
+import { environment } from 'environments/environment.dev';
 
 @Injectable({
   providedIn: 'root'
