+++ src/app/loans/loans-view/loan-documents-tab/loan-documents-tab.component.ts
@@ -3,7 +3,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 
 /** Custom Services */
-import { environment } from 'environments/environment';
+import { environment } from 'environments/environment.dev';
 import { LoansService } from 'app/loans/loans.service';
 import { SettingsService } from 'app/settings/settings.service';
