+++ src/app/savings/savings-account-view/savings-account-view.component.ts
@@ -17,7 +17,7 @@ import { Currency } from 'app/shared/models/general.model';
 import { TranslateService } from '@ngx-translate/core';
 
 /** Environment Configuration */
-import { environment } from 'environments/environment';
+import { environment } from 'environments/environment.dev';
 
 /**
  * Savings Account View Component
