+++ src/app/shared/footer/footer.component.ts
@@ -9,7 +9,7 @@ import { SystemService } from 'app/system/system.service';
 import { VersionService } from 'app/system/version.service';
 
 /** Environment Configuration */
-import { environment } from 'environments/environment';
+import { environment } from 'environments/environment.dev';
 import { Subscription } from 'rxjs';
 
 /**
