+++ src/app/shared/notifications-tray/notifications-tray.component.ts
@@ -6,7 +6,7 @@ import { forkJoin } from 'rxjs';
 
 /** Custom Services */
 import { NotificationsService } from 'app/notifications/notifications.service';
-import { environment } from 'environments/environment';
+import { environment } from 'environments/environment.dev';
 
 /**
  * Notifications Tray Component
