+++ src/main.ts
@@ -8,7 +8,7 @@ import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
-import { environment } from './environments/environment';
+import { environment } from './environments/environment.dev';
 
 if (environment.production) {
   enableProdMode();
